+++ client/src/app/budget/budgets.tsx
@@ -1,44 +1,32 @@
 import MonthIterator from './month-iterator';
 import React from 'react';
 import AddButton from '@/components/add-button';
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { BudgetGroup, getBudgets, getBudgetsForMonth, getBudgetsForGroup } from '@/lib/budgets';
 import BudgetCards from './budget-cards';
 import AddBudget from './add-budget';
-import { getTransactions, getTransactionsForMonth } from '@/lib/transactions';
+import { getTransactionsForMonth } from '@/lib/transactions';
 import BudgetHeader from './budget-header';
 import { type Budget } from '@/types/budget';
 import BudgetTotalCard from './budget-total-card';
 import { initMonth } from '@/lib/utils';
 import { type Transaction } from '@/types/transaction';
-import request from '@/lib/request';
-import { type AxiosResponse } from 'axios';
+import request, { translateAxiosError } from '@/lib/request';
+import { type AxiosResponse, type AxiosError } from 'axios';
 import { defaultGuid } from '@/types/user';
 import ResponsiveButton from '@/components/responsive-button';
-import AlertBanner from '@/components/alert-banner';
 import Unbudgets from './unbudgets';
+import { useBudgetsQuery, useTransactionsQuery } from '@/lib/query';
+import { useToast } from '@/components/ui/use-toast';
 
 const Budgets = (): JSX.Element => {
   const [date, setDate] = React.useState<Date>(initMonth());
-  const [alert, setAlert] = React.useState('');
 
-  const budgetsQuery = useQuery({
-    queryKey: ['budgets', { date }],
-    queryFn: async () => {
-      const response = await getBudgets(date);
-      return response;
-    },
-  });
-
-  const transactionsQuery = useQuery({
-    queryKey: ['transactions'],
-    queryFn: async () => {
-      const response = await getTransactions();
-      return response;
-    },
-  });
+  const budgetsQuery = useBudgetsQuery(date);
+  const transactionsQuery = useTransactionsQuery();
 
   const queryClient = useQueryClient();
+  const { toast } = useToast();
   const doCopyBudget = useMutation({
     mutationFn: async (newBudgets: Budget[]) => {
       return await request({
@@ -50,12 +38,18 @@ const Budgets = (): JSX.Element => {
     onSuccess: async () => {
       await queryClient.invalidateQueries({ queryKey: ['budgets'] });
     },
+    onError: (error: AxiosError) => {
+      toast({
+        variant: 'destructive',
+        title: 'Uh oh! Something went wrong.',
+        description: translateAxiosError(error),
+      });
+    },
   });
 
   return (
     <div className="flex w-screen max-w-screen-2xl flex-row justify-center space-x-2">
       <div className="w-3/4 flex-grow space-y-2">
-        <AlertBanner alert={alert} setAlert={setAlert} />
         <div className="grid w-full grid-cols-3">
           <div />
           <div className="justify-self-center">
@@ -81,11 +75,17 @@ const Budgets = (): JSX.Element => {
                         });
                         doCopyBudget.mutate(res.data as Budget[]);
                       } else {
-                        setAlert('Last month has no budget!');
+                        toast({
+                          variant: 'destructive',
+                          description: 'Last month has no budget!',
+                        });
                       }
                     })
                     .catch(() => {
-                      setAlert("There was an error copying last month's budget.");
+                      toast({
+                        variant: 'destructive',
+                        description: "There was an error copying last month's budget.",
+                      });
                     });
                 }}
               >
